Turn inspections up to 11
OK, so let's get this out of the way, there's some actual changes mixed
in here too. I'm really sorry:
 - Turtles can now not be renamed with unnamed item tags (previously it
   would clear the name, this seemed a little unideal).
 - commands.getBlock(s)Data will also include NBT.

Now, onto the horror story which is these inspection changes:
 - Make a lot of methods static
 - Typo fixes
 - Make utility classes final + private constructor
 - Lots of reformatting (ifs -> ternary, invert control flow, etc...)
 - ???
 - Profit!

I'm so going to regret this - can pretty much guarantee this is going to
break something.

+++ src/main/java/dan200/computercraft/shared/util/ColourTracker.java
@@ -43,7 +43,7 @@ public class ColourTracker
         int avgG = totalG / count;
         int avgB = totalB / count;
 
-        float avgTotal = (float) total / (float) count;
+        float avgTotal = (float) total / count;
         float avgMax = Math.max( avgR, Math.max( avgG, avgB ) );
         avgR = (int) (avgR * avgTotal / avgMax);
         avgG = (int) (avgG * avgTotal / avgMax);
